// src/app/parametros-atendimento-paciente/shared/model/parametros-atendimento-paciente.form.ts
export class ParametrosAtendimentoPacienteFORM {

    quantidadeParcelas: number;
    tempoMesesGeracaoAutomaticaHorariosAtendimento: number;
    intervaloMinutosEntreAtendimentos: number;
    intervaloDiasEntrePrimeiraConsultaRetorno: number;
    intervaloDiasEntreConsultaRetorno: number;
    intervaloDiasEntreRetornoConsulta: number;
    precoConsulta: string;

    tempoMesesAtualizarQuestionarioFrequenciaAlimentar: number;
	tempoMesesAtualizarHistoricoSocial: number;
	tempoMesesAtualizarHistoricoPatologiaFamiliares: number;
	tempoMesesAtualizarHistoricoAtividadeFisica: number;
	tempoMesesAtualizarHistoricoAlimentar: number;
	tempoMesesAtualizarHistoricoFrequenciaAlimentar: number;
}